+++ src/quantum/perfherder.jsx
@@ -15,7 +15,7 @@ import {
   timeMonth,
 } from 'd3';
 import { stringify } from 'query-string';
-import { withErrorBoundary } from '../vendor/errors';
+import withErrorBoundary from '../vendor/errors';
 import Widget from './widget';
 import SETTINGS from '../settings';
 
@@ -24,8 +24,8 @@ const tickCount = 4;
 class PerfherderWidget extends React.Component {
   state = {};
 
-  componentDidMount() {
-    this.fetch();
+  async componentDidMount() {
+    await this.fetch();
   }
 
   async fetch() {
