+++ src/telemetry/graph.jsx
@@ -5,7 +5,8 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import { stringify } from 'query-string';
 import SETTINGS from '../settings';
-import { Exception, withErrorBoundary } from '../vendor/errors';
+import withErrorBoundary from '../vendor/errors';
+import { Exception } from '../vendor/logs';
 
 class TelemetryContainer extends React.Component {
   async componentDidMount() {
